+++ Sensor/Proximity/Infrared/single.h
@@ -8,8 +8,6 @@
 #ifndef SENSOR_PROXIMITY_INFRARED_SINGLE_H
 #define SENSOR_PROXIMITY_INFRARED_SINGLE_H
 
-#include <stdint.h>
-
-uint8_t sensor_infrared_single_read (uint8_t);
+uint16_t sensor_infrared_single_read (uint8_t);
 
 #endif
